Move naming validator into owned naming slice root
