+++ calculogic-validator/bin/calculogic-validate-naming.mjs
@@ -5,7 +5,7 @@ import {
   summarizeFindings,
   listNamingValidatorScopes,
   getScopeProfile,
-} from '../src/naming/naming-validator.host.mjs';
+} from '../naming/src/naming-validator.host.mjs';
 import { resolveRepositoryRoot } from '../src/core/repository-root.logic.mjs';
 import { loadValidatorConfigFromFile } from '../src/core/config/validator-config.logic.mjs';
 import {
@@ -181,5 +181,5 @@ const report = {
   findings,
 };
 
-console.log(JSON.stringify(report, null, 2));
-process.exit(deriveExitCodeFromFindings(findings, { strict: parsed.strict }));
+process.stdout.write(`${JSON.stringify(report, null, 2)}\n`);
+process.exitCode = deriveExitCodeFromFindings(findings, { strict: parsed.strict });
