+++ calculogic-validator/doc/Audits/final-state-validator-shim-verification-audit.md
@@ -63,7 +63,7 @@ Key final-state conclusion: detector output now aligns cleanly with the manually
 ### 4.2 Intentional canonical/public pass-through entrypoints (non-debt, 3)
 
 1. `calculogic-validator/src/index.mjs` (public barrel entrypoint)
-2. `calculogic-validator/src/naming/naming-validator.host.mjs` (canonical host→wiring boundary)
+2. `calculogic-validator/naming/src/naming-validator.host.mjs` (canonical host→wiring boundary)
 3. `calculogic-validator/tree/src/tree-structure-advisor.host.mjs` (canonical host→wiring boundary)
 
 ## 5. Current tree validator comparison
@@ -77,7 +77,7 @@ Key final-state conclusion: detector output now aligns cleanly with the manually
 ### 5.2 Intentional pass-through carveouts correctly unflagged
 
 - `calculogic-validator/src/index.mjs` remains unflagged (public barrel carveout preserved).
-- `calculogic-validator/src/naming/naming-validator.host.mjs` remains unflagged (canonical host boundary carveout preserved).
+- `calculogic-validator/naming/src/naming-validator.host.mjs` remains unflagged (canonical host boundary carveout preserved).
 - `calculogic-validator/tree/src/tree-structure-advisor.host.mjs` remains unflagged (canonical host boundary carveout preserved).
 
 ### 5.3 False positive / ambiguous / missed classification
