+++ calculogic-validator/naming/src/naming-validator.contracts.mjs
