+++ calculogic-validator/naming/src/registries/_builtin/case-rules.registry.json
