+++ calculogic-validator/naming/src/registries/_builtin/categories.registry.json
