+++ calculogic-validator/naming/src/registries/_builtin/roles.registry.json
