+++ calculogic-validator/naming/src/registries/_builtin/special-cases.registry.json
