+++ calculogic-validator/naming/src/registries/_builtin/walk-exclusions.registry.json
