+++ calculogic-validator/naming/src/registries/_custom/roles.registry.custom.json
