+++ calculogic-validator/naming/src/registries/naming-special-case-rules.registry.logic.mjs
