+++ calculogic-validator/naming/src/registries/naming-walk-exclusions.registry.logic.mjs
