+++ calculogic-validator/naming/src/rules/naming-rule-check-hyphen-role-ambiguity.logic.mjs
