+++ calculogic-validator/naming/src/rules/naming-rule-check-role.logic.mjs
