+++ calculogic-validator/naming/src/rules/naming-rule-check-semantic-case.logic.mjs
