+++ calculogic-validator/naming/src/rules/naming-rule-classify-special-case.logic.mjs
