+++ calculogic-validator/naming/src/rules/naming-rule-parse-canonical.logic.mjs
