+++ calculogic-validator/naming/test/naming-case-rules-registry.test.mjs
@@ -1,8 +1,8 @@
 import test from 'node:test';
 import assert from 'node:assert/strict';
-import caseRulesRegistry from '../src/naming/registries/_builtin/case-rules.registry.json' with { type: 'json' };
-import { getSemanticNameCaseRule } from '../src/naming/rules/naming-rule-check-semantic-case.logic.mjs';
-import { isCanonicalSemanticName } from '../src/naming/rules/naming-rule-check-semantic-case.logic.mjs';
+import caseRulesRegistry from '../src/registries/_builtin/case-rules.registry.json' with { type: 'json' };
+import { getSemanticNameCaseRule } from '../src/rules/naming-rule-check-semantic-case.logic.mjs';
+import { isCanonicalSemanticName } from '../src/rules/naming-rule-check-semantic-case.logic.mjs';
 
 test('semantic-name case rule runtime is sourced from builtin registry style', () => {
   const semanticCaseRule = getSemanticNameCaseRule();
