+++ calculogic-validator/naming/test/naming-walk-exclusions-registry.test.mjs
@@ -3,11 +3,11 @@ import assert from 'node:assert/strict';
 import fs from 'node:fs';
 import os from 'node:os';
 import path from 'node:path';
-import { collectRepositoryPaths } from '../src/naming/naming-validator.host.mjs';
+import { collectRepositoryPaths } from '../src/naming-validator.host.mjs';
 import {
   BUILTIN_WALK_EXCLUSIONS_REGISTRY_PATH,
   getBuiltinWalkExclusions,
-} from '../src/naming/registries/naming-walk-exclusions.registry.logic.mjs';
+} from '../src/registries/naming-walk-exclusions.registry.logic.mjs';
 
 const writeFile = (rootDirectory, relativePath) => {
   const absolutePath = path.join(rootDirectory, relativePath);
