+++ calculogic-validator/scripts/validator-health-check.host.mjs
@@ -3,7 +3,7 @@ import {
   runNamingValidator,
   summarizeFindings,
   getScopeProfile,
-} from '../src/naming/naming-validator.host.mjs';
+} from '../naming/src/naming-validator.host.mjs';
 import fs from 'node:fs';
 import { resolveRepositoryRoot } from '../src/core/repository-root.logic.mjs';
