+++ calculogic-validator/src/index.mjs
@@ -1,6 +1,6 @@
 export * from './core/validator-runner.logic.mjs';
 export * from './core/validator-registry.knowledge.mjs';
 export * from './core/validator-report.contracts.mjs';
-export * as naming from './naming/naming-validator.host.mjs';
+export * as naming from '../naming/src/naming-validator.host.mjs';
 
 export * as treeStructureAdvisor from '../tree/src/tree-structure-advisor.host.mjs';
